requirements.txt
python-magic-bin==0.4.14
dearpygui==1.8.0
DearPyGui-DragAndDrop==1.0.0
tabulate==0.9.0
pydub==0.25.1
